t/04_fail.t: prevent permission test on MSWin32
In addition to POSIX systems with uid 0, where any file can
be read, don't test on Windows either, since it seems the
permissions aren't working as expected.

Since the tests are for code completion only and intended to
catch accidently added failures in error path, it's fine to
test them on less runs then the general operaion tests.

Signed-off-by: Jens Rehsack <sno@netbsd.org>