use eval+require rather than use_ok to test for CPAN::Meta to fix plan
A skip_all plan is incompatible with having any tests.  If CPAN::Meta
was not available, the use_ok would omit a failed TODO test, but then an
incompatible skip_all plan.

Rather than using use_ok, just use a require in an eval block to check
for CPAN::Meta.